application state has been moved to database

+++ src/main/java/ru/trofimov/app/operations/impl/AccountCreateOperation.java
@@ -22,7 +22,7 @@ public class AccountCreateOperation implements Operation {
     public void execute(Scanner scanner) {
         System.out.println("Enter the user id for which to create an account: ");
 
-        int userId = scanner.nextInt();
+        Long userId = scanner.nextLong();
         scanner.nextLine();
         User user = accountService.createAccount(userId);
         System.out.println("New account created with ID: " + userId + " for user: " + user.getLogin());
